Move polus folder from src to root
