+++ polus/trajectories/commons.py
