+++ polus/trajectories/globals.py
